Separate sites selectors.
They all lived in a massive file, which made it very hard to code-split
effectively.
Each of them lives in its own module now, which can be imported
separately. To ease migration, there's still an index file which
imports and reexports them all.
