+++ client/state/sites/selectors/can-current-user-upgrade-site.js
@@ -1,0 +1,27 @@
+/**
+ * Internal dependencies
+ */
+import { isCurrentUserCurrentPlanOwner } from 'state/sites/plans/selectors';
+import { getSelectedSiteId } from 'state/ui/selectors';
+import canCurrentUser from 'state/selectors/can-current-user';
+import isCurrentPlanPaid from './is-current-plan-paid';
+
+/**
+ * Returns true if current user can purchase upgrades for this site
+ *
+ * @param  {Object}   state  Global state tree
+ * @param  {Number}   siteId Site ID
+ * @return {?Boolean}        Whether site is previewable
+ */
+export default function canCurrentUserUpgradeSite( state, siteId = null ) {
+	if ( ! siteId ) {
+		siteId = getSelectedSiteId( state );
+	}
+	const canUserManageOptions = canCurrentUser( state, siteId, 'manage_options' );
+	if ( ! canUserManageOptions ) {
+		return false;
+	}
+
+	const isPaid = isCurrentPlanPaid( state, siteId );
+	return ! isPaid || isCurrentUserCurrentPlanOwner( state, siteId );
+}
