+++ client/state/sites/selectors/is-current-plan-paid.js
@@ -1,0 +1,21 @@
+/**
+ * Internal dependencies
+ */
+import getSitePlan from './get-site-plan';
+
+/**
+ * Returns true if the current site plan is a paid one
+ *
+ * @param  {Object}   state         Global state tree
+ * @param  {Number}   siteId        Site ID
+ * @return {?Boolean}               Whether the current plan is paid
+ */
+export default function isCurrentPlanPaid( state, siteId ) {
+	const sitePlan = getSitePlan( state, siteId );
+
+	if ( ! sitePlan ) {
+		return null;
+	}
+
+	return sitePlan.product_id !== 1 && sitePlan.product_id !== 2002;
+}
