+++ client/state/sites/selectors/is-current-site-plan.js
@@ -1,0 +1,26 @@
+/**
+ * Internal dependencies
+ */
+import getSitePlan from './get-site-plan';
+
+/**
+ * Returns true if site is currently subscribed to supplied plan and false otherwise.
+ *
+ * @param  {Object}   state         Global state tree
+ * @param  {Number}   siteId        Site ID
+ * @param  {Number}   planProductId Plan product_id
+ * @return {?Boolean}               Whether site's plan matches supplied plan
+ */
+export default function isCurrentSitePlan( state, siteId, planProductId ) {
+	if ( planProductId === undefined ) {
+		return null;
+	}
+
+	const sitePlan = getSitePlan( state, siteId );
+
+	if ( ! sitePlan ) {
+		return null;
+	}
+
+	return sitePlan.product_id === planProductId;
+}
