+++ client/state/sites/selectors/is-jetpack-site.js
@@ -1,0 +1,21 @@
+/**
+ * Internal dependencies
+ */
+import getRawSite from 'state/selectors/get-raw-site';
+
+/**
+ * Returns true if site is a Jetpack site, false if the site is hosted on
+ * WordPress.com, or null if the site is unknown.
+ *
+ * @param  {Object}   state  Global state tree
+ * @param  {Number}   siteId Site ID
+ * @return {?Boolean}        Whether site is a Jetpack site
+ */
+export default function isJetpackSite( state, siteId ) {
+	const site = getRawSite( state, siteId );
+	if ( ! site ) {
+		return null;
+	}
+
+	return site.jetpack;
+}
