+++ client/state/sites/selectors/is-requesting-sites.js
@@ -1,0 +1,8 @@
+/**
+ * Returns true if we are requesting all sites.
+ * @param {Object}    state  Global state tree
+ * @return {Boolean}        Request State
+ */
+export default function isRequestingSites( state ) {
+	return !! state.sites.requestingAll;
+}
